backend/requirements.txt
django
djangorestframework
drf-spectacular
django-cors-headers
django-spa
django-filter
pre-commit
